update docs and image example
