+++ bencher/example/example_docs.py
@@ -15,4 +15,4 @@ if __name__ == "__main__":
     runner.add(example_meta_float)
     runner.add(example_meta_levels)
 
-    runner.run(level=2, grouped=True, show=True)
+    runner.run(level=2, grouped=True, show=True,cache_results=False)
